Started implementation
Currently the cd() function just splits the newPath string in to an array using the split() method, then prints each element of the array. I did this to test how the split() function works in Java, and more specifically to see how it behaves when the first character of the string to split is the given regular expression. This behavior is important to understand for the case where the function is passed an absolute pathname