Finished declaring variables
I believe that I have finished declaring all variables necessary to implement the cd() function. Along with the array of strings created by applying split("/") to the newPath string argument, I have also declared+initialized a list which holds the current path called pathList. 

I chose for this to be a list, rather than an array, because the length of the current path will change during execution of the function. 

The code still doesn't actually do anything besides print a test statement.