Finished implementation
Finished implementing all required functionality. It was a bit trickier to update the path String after setting the pathList ArrayList than I had expected since I ran in to a few complications properly inserting the "/" characters. 

The solution I ended up finding was to create a StringBuilder variable called updatedPath. Then I used a for loop to append the "/" and an element of the pathList ArrayList to the new updatedPath StringBuilder, for each element in pathList. Finally, I converted this updatedPath StringBuilder to a String, and set the path variable to this value.

I am sure there was a simpler approach using some of the methods of the ArrayList class, but I was having a few troubles and this approach worked. Passes all 4/4 tests on testdome.com.